Factor out common ECS logic into helpers
