+++ python_modules/libraries/dagster-aws/dagster_aws/ecs/__init__.py
@@ -1,1 +1,2 @@
-from .launcher import EcsEventualConsistencyTimeout, EcsRunLauncher
+from .launcher import EcsRunLauncher
+from .tasks import EcsEventualConsistencyTimeout
